Polish init progress rendering and embeddings bootstrap flow
